requirements.txt
discord-py
python-dotenv
tqdm
requests
aiohttp